Batch Update with Optimistic Locking control.
Change made to be able to use batch update when optimistic locking
control is in use.

All test passed, but there is a change in the type of exception is
thrown, in many cases StaleObjectStateException was no thrown anymore
but StaleStateException. In other case the StaleObjectStateException was
still in use but didn't bring any info of the EntityName and Id.

https://nhibernate.jira.com/browse/NH-3771