Flush child sessions when parent is flushed.
For years I used GeneratedBy.Assigned() with a incremental PK for audit logging.  As soon as Save is called it issues an INSERT to the database, so it didn't matter if Flush was called on the session.  I switched to guid.comb and now Save does not issue an INSERT (which is the correct behavior) but since the child session isn't flushed my audit record is never written.