Make OfType work better with polymorphic mappings
Addresses NH-3845.  The ProcessOfType Linq visitor class now handles interfaces and base classes by checking which mapped types that could be in the result set implement the class, then filtering accordingly.  If no results are possible, any remaining filters are removed from the query tree to avoid problems with invalid property names.

Because the HQL tree for OfType can now mimic the ".class IN" construct, the In method of HqlTreeBuilder was inadequate to express the tree.  Allowing an array of nodes to be passed as the source solves this problem.