Improve source-type determination in OfType
Continues to address NH-3845.  Because OfType is one of the result operators removed from its subquery and attached to the main query, it is not always clear during the ProcessOfType.Process method which query source should be used to retrieve the .class property of the entity.  Sometimes the query type is in the QueryModel.MainFromClause, and sometimes it is in the QueryModel.SelectClause.  It may be elsewhere, also.  For queries in which OfType will match some but not necessarily all results, attempt to determine and use the query name as an identifier rather than using the source expression, for example.  Doing so simplifies this particular Process method while still allowing query source determination later.