Batch Update with Optimistic Lock
This is a suggestion for batch update when optimistic control. Now, when
version control is active, batch update is not done, see the fixture.
This is a down side in performance.

We made the change, and run the tests. The diference is that
StaleStateException is thrown instead of StaleObjectStateException, in
which case we loose the information regarding the entity and it's id
that generate the exception.

In our application, this is not a big problem, compare with the
possibility of doing a batch operation instead a single one.

We have been looking at the source code, and will be a possibility to
still get a class and a group of id's on the batcher to generate a more
accurate exception.

Please, let me know if this will be of intereset, so i can go deeper in
that aproach, or if you see another aproach that you will suggest.