Resharper settings. Revert to USE_SPACES for alignment.
After some discussion between myself and Alexander, we've agreed to go back
to USE_SPACES. To summarize the arguments:

USE_SPACES and OPTIMAL_FILL both have the drawback of mixing tabs
and spaces, but OPTIMAL_FILL have the additional problem of giving inconsistent
results with different tab width settings. It appears impractical to ensure
the same tab width is used in ALL tools, also in web based tools.

USE_TABS is useless because it doesn't allow correct indenting.

Switching to using space only, for both indenting and alignment, would give
consistent result, but would involve a complete reformat of all files, which
we don't want to do again.

So, it remains that tabs for indent plus USE_SPACES for alignment is the
least bad option that's practical for us.