AbstractCollectionPersister: Include the table when generating column aliases.
This modifies the calls to Column.GetAlias() in two places to also include
the table parameter, so that the generated alias will be unique in more cases.
It matches what is currently used in Hibernate code.

This fixes the failure of SubselectFetchFixture.ManyToManyCriteriaJoin on
SqlServerCe that was introduced by the modifications to Column.GetAlias()
done in 013b5da27301.