Fix one-to-one with property-ref stack overflow
Fix key type mismatch in EntityUniqueKey
EntityUniqueKey was assuming the value to be always semi-resolved (to be
the identifier instead of the entity for an entity type). It was always
computing the value hashcode with the semi-resolved type, which is the
identifier type for an entity.
But this has changed with #1492, which caches unique keys also for
already fully loaded entities. The value hashcode must then be computed
with the entity type, which accounts for proxies.
Otherwise the proxy loading may get triggered while the code is already
loading the proxy, which results in a stack overflow.

Add test case adapted from Michael Estermann repro

Fix #1645

Co-authored-by: Michael Estermann <michael.estermann@bbv.ch>