Handle multi-queries support in FutureBatch  (#1666)
The previous Future implementation was letting all implementors handling
the lack of multi-queries support themselves, leading to many
discrepancies in behavior with a data provider which did not support
multi-queries:

 * Accessing the future value was causing the query to be executed at
   each access.
 * Only the accessed future was executed, not all the other defined
   futures. This is the cause of #1663.
 * Futures having an ExecuteOnEval were not executing it, leading to
   NH-3850 bugs.

Now FutureBatch handles itself the multi-queries support or lack of
support.

NH-3850 was not fixed for data providers which did not support
multi-queries. But this was undetected because its tests were executed
on too few databases, due to them requiring the support of
DateTimeOffset.