Fix cascade + one-to-one failing on null no-proxy association
This is a lightweight fix which does not change the no-proxy principles
at the root of the bug.
The no-proxy is currently implemented as a special case in the lazy
property interceptor, not flagged as a lazy property. Instead, on parent
entity load, it hides, in the parent loaded state for the no-proxy, a
proxy (!) flagged for unwrapping the associated entity it proxifies.
This hiden unwrapping proxy supports proxifying  non-existent entities.
The property interceptor does the unwrapping when accessing it. But when
the proxy unwraps to null, this causes the loaded state to have a
reference where it should be indeed null, causing dirty checking issues.

A better fix would be to flag the property as lazy for no-proxy
one-to-one associations, allowing to get in the loadedState the
"uninitializedProperty" instance, allowing the dirty checking mechanism
to know that this part of state is unknown (till loading it, at which
point it should be adjusted to its actual value).
But this causes the property interceptor to trigger initialization of
any other lazy properties when accessing a one-to-one. Avoiding this
would require to implement some lazy property fetch group semantic.

Fixes #1719

Co-authored-by: Alexander Zaytsev <hazzik@gmail.com>