Move WcfOperationSessionContext into a separate project
NHibernate currently references the System.ServiceModel.Primitives package, which along with it's own dependencies, adds ~4 MB to a published .NET Core application. The reference is only used to provide WcfOperationSessionContext. This PR creates a new project and NuGet package, NHibernate.Wcf, and moves WcfOperationSessionContext and the System.ServiceModel.Primitives reference accordingly, allowing users to chose when to opt-in to the additional dependencies.