Fix attempt of static proxies to call base method for abstract classes (#1884)
Previously it was failing PEVerify similarly to #1728 but for abstract classes (this is possible in case of a polymorphic entities). Unlike interfaces the abstract base class can go into a situation when lazy initializer is not yet available, eg. code in a constructor.

Also test the explicit interface case, check access into constructor, and check proxy on type with
interfaces.

And correctly call explicit interface methods of base type